+++ server/aws-lsp-codewhisperer/src/language-server/inline-completion/codeWhispererServer.ts
@@ -225,6 +225,8 @@ export const CodewhispererServerFactory =
                     const maxResults = isAutomaticLspTriggerKind ? 1 : 5
                     const selectionRange = params.context.selectedCompletionInfo?.range
 
+                    const startPreprocessTimestamp = Date.now()
+
                     // For Jupyter Notebook in VSC, the language server does not have access to
                     // its internal states including current active cell index, etc
                     // we rely on VSC to calculate file context
@@ -256,7 +258,7 @@ export const CodewhispererServerFactory =
                     const previousSession = completionSessionManager.getPreviousSession()
                     // Only refer to decisions in the past 2 mins
                     const previousDecisionForClassifier =
-                        previousSession && performance.now() - previousSession.decisionMadeTimestamp <= 2 * 60 * 1000
+                        previousSession && Date.now() - previousSession.decisionMadeTimestamp <= 2 * 60 * 1000
                             ? previousSession.getAggregatedUserTriggerDecision()
                             : undefined
                     let ideCategory: string | undefined = ''
@@ -357,6 +359,7 @@ export const CodewhispererServerFactory =
 
                     const newSession = completionSessionManager.createSession({
                         document: textDocument,
+                        startPreprocessTimestamp: startPreprocessTimestamp,
                         startPosition: params.position,
                         triggerType: isAutomaticLspTriggerKind ? 'AutoTrigger' : 'OnDemand',
                         language: fileContext.programmingLanguage.languageName as CodewhispererLanguage,
@@ -425,8 +428,8 @@ export const CodewhispererServerFactory =
                 session.suggestions = suggestionResponse.suggestions
                 session.responseContext = suggestionResponse.responseContext
                 session.codewhispererSessionId = suggestionResponse.responseContext.codewhispererSessionId
-                session.timeToFirstRecommendation = new Date().getTime() - session.startTime
-                session.suggestionType = suggestionResponse.suggestionType
+                session.setTimeToFirstRecommendation()
+                session.predictionType = SuggestionType.COMPLETION
             } else {
                 session.suggestions = [...session.suggestions, ...suggestionResponse.suggestions]
             }
@@ -454,7 +457,7 @@ export const CodewhispererServerFactory =
 
             // session was closed by user already made decisions consequent completion request before new paginated API response was received
             if (
-                session.suggestionType !== SuggestionType.EDIT && // TODO: this is a shorterm fix to allow Edits tabtabtab experience, however the real solution is to manage such sessions correctly
+                session.predictionType !== SuggestionType.EDIT && // TODO: this is a shorterm fix to allow Edits tabtabtab experience, however the real solution is to manage such sessions correctly
                 (session.state === 'CLOSED' || session.state === 'DISCARD')
             ) {
                 return EMPTY_RESULT
@@ -640,7 +643,7 @@ export const CodewhispererServerFactory =
             let deletedLengthForEdits = 0
             if (acceptedSuggestion) {
                 codePercentageTracker.countSuccess(session.language)
-                if (session.suggestionType === SuggestionType.EDIT && acceptedSuggestion.content) {
+                if (session.predictionType === SuggestionType.EDIT && acceptedSuggestion.content) {
                     // [acceptedSuggestion.insertText] will be undefined for NEP suggestion. Use [acceptedSuggestion.content] instead.
                     // Since [acceptedSuggestion.content] is in the form of a diff, transform the content into addedCharacters and deletedCharacters.
                     const { addedLines, deletedLines } = getAddedAndDeletedLines(acceptedSuggestion.content)
@@ -843,9 +846,9 @@ export const CodewhispererServerFactory =
 
             // Record last user modification time for any document
             if (lastUserModificationTime) {
-                timeSinceLastUserModification = new Date().getTime() - lastUserModificationTime
+                timeSinceLastUserModification = Date.now() - lastUserModificationTime
             }
-            lastUserModificationTime = new Date().getTime()
+            lastUserModificationTime = Date.now()
 
             documentChangedListener.onDocumentChanged(p)
             editCompletionHandler.documentChanged()