+++ server/aws-lsp-codewhisperer/src/language-server/inline-completion/session/sessionManager.test.ts
@@ -28,6 +28,7 @@ describe('CodeWhispererSession', function () {
 
     const data: SessionData = {
         document: TextDocument.create('file:///rightContext.cs', 'csharp', 1, HELLO_WORLD_IN_CSHARP),
+        startPreprocessTimestamp: 0,
         startPosition: { line: 0, character: 0 },
         triggerType: 'OnDemand',
         language: 'csharp',
@@ -510,6 +511,7 @@ describe('SessionManager', function () {
     }
     const data: SessionData = {
         document: TextDocument.create('file:///rightContext.cs', 'csharp', 1, HELLO_WORLD_IN_CSHARP),
+        startPreprocessTimestamp: 0,
         startPosition: { line: 0, character: 0 },
         triggerType: 'OnDemand',
         language: 'csharp',